fix #394 rename discover module to discovery
