+++ discovery/src/main/java/com/alibaba/fescar/discovery/loadbalance/RoundRobinLoadBalance.java
@@ -14,25 +14,18 @@
  *  limitations under the License.
  */
 
-package com.alibaba.fescar.discover.loadbalance;
+package com.alibaba.fescar.discovery.loadbalance;
 
 import java.util.List;
 
 /**
- * The interface Load balance.
- *
- * @author: jimin.jm @alibaba-inc.com
- * @date 2019 /02/12
+ * @author: jimin.jm@alibaba-inc.com
+ * @date 2019/02/12
  */
-public interface LoadBalance {
-
-    /**
-     * Select t.
-     *
-     * @param <T>      the type parameter
-     * @param invokers the invokers
-     * @return the t
-     * @throws Exception the exception
-     */
-    <T> T select(List<T> invokers) throws Exception;
+public class RoundRobinLoadBalance extends AbstractLoadBalance {
+    @Override
+    protected <T> T doSelect(List<T> invokers) {
+        //todo
+        return invokers.get(0);
+    }
 }
